Split native link cache helper module
